HMW-315 Fixed loading spinner and resetting clustering.
