+++ app/client/src/types/index.ts
@@ -196,6 +196,7 @@ interface ServicesData {
   waterQualityPortal: {
     resultSearch: string;
     userInterface: string;
+    monitoringLocation: string;
   };
 }
